Added important functionalities to utils file

+++ anchor/programs/anchor/src/constants.rs
@@ -7,9 +7,9 @@ pub const SEED_SOL_ACC: &[u8] = b"sol_acc";
 
 #[constant]
 pub const MINT_DECIMALS: u8 = 9;
-pub const LIQUIDATION_BONUS: u8 = 10;
-pub const MIN_HEALTH_FACTOR: u8 = 1;
-pub const LIQUIDATION_THRESHOLD: u8 = 50;
+pub const LIQUIDATION_BONUS: u64 = 10;
+pub const MIN_HEALTH_FACTOR: u64 = 1;
+pub const LIQUIDATION_THRESHOLD: u64 = 50;
 
 pub const FEED_ID: &str = "0xef0d8b6fda2ceba41da15d4095d1da392a0d2f8ed0c6c7bc0f4cfac8c280b56d";
 pub const MAXIMUM_AGE_PRICE_FEED: u64 = 60;
